move validateOrderItemForm to helpers and fix manifestProvider
