+++ app/pages/sections/catalogue-solutions/order-item/controller.js
@@ -1,11 +1,9 @@
-/* eslint-disable no-restricted-globals */
 import { getData, postData, putData } from 'buying-catalogue-library';
 import { getContext, getErrorContext } from './contextCreator';
 import { logger } from '../../../../logger';
 import { getEndpoint } from '../../../../endpoints';
 import commonManifest from './commonManifest.json';
-import { getSelectedPriceManifest } from './manifestProvider';
-import { getDateErrors } from '../../../../helpers/controllers/getDateErrors';
+import { getSelectedPriceManifest } from '../../../../helpers/controllers/manifestProvider';
 import { extractDate } from '../../../../helpers/controllers/extractDate';
 
 export const getOrderItem = async ({ orderId, orderItemId, accessToken }) => {
@@ -92,80 +90,6 @@ export const getOrderItemErrorPageContext = (params) => {
   };
 
   return getErrorContext(updatedParams);
-};
-
-export const validateOrderItemForm = ({ data, selectedPrice }) => {
-  const errors = [];
-  const selectedPriceManifest = getSelectedPriceManifest({
-    provisioningType: selectedPrice.provisioningType,
-    type: selectedPrice.type,
-  });
-
-  if (selectedPriceManifest.questions.deliveryDate) {
-    const deliveryDateError = getDateErrors('deliveryDate', data);
-    if (deliveryDateError) {
-      errors.push(deliveryDateError);
-    }
-  }
-
-  if (selectedPriceManifest.questions.quantity) {
-    if (!data.quantity || data.quantity.trim().length === 0) {
-      errors.push({
-        field: 'Quantity',
-        id: 'QuantityRequired',
-      });
-    } else if (isNaN(data.quantity)) {
-      errors.push({
-        field: 'Quantity',
-        id: 'QuantityMustBeANumber',
-      });
-    } else if (data.quantity.indexOf('.') !== -1) {
-      errors.push({
-        field: 'Quantity',
-        id: 'QuantityInvalid',
-      });
-    } else if (data.quantity > 2147483646) {
-      errors.push({
-        field: 'Quantity',
-        id: 'QuantityLessThanMax',
-      });
-    }
-  }
-
-  if (selectedPriceManifest.questions.selectEstimationPeriod) {
-    if (!data.selectEstimationPeriod) {
-      errors.push({
-        field: 'SelectEstimationPeriod',
-        id: 'EstimationPeriodRequired',
-      });
-    }
-  }
-
-  if (selectedPriceManifest.addPriceTable.cellInfo.price.question) {
-    if (!data.price || data.price.trim().length === 0) {
-      errors.push({
-        field: 'Price',
-        id: 'PriceRequired',
-      });
-    } else if (isNaN(data.price)) {
-      errors.push({
-        field: 'Price',
-        id: 'PriceMustBeANumber',
-      });
-    } else if (data.price.includes('.') && data.price.split('.')[1].length > 3) {
-      errors.push({
-        field: 'Price',
-        id: 'PriceMoreThan3dp',
-      });
-    } else if (parseFloat(data.price) > 999999999999999.999) {
-      errors.push({
-        field: 'Price',
-        id: 'PriceLessThanMax',
-      });
-    }
-  }
-
-  return errors;
 };
 
 export const getSolution = async ({ solutionId, accessToken }) => {
